--- tsconfig.test.json
@@ -1,27 +1,0 @@
-{
-  "compilerOptions": {
-    "target": "esnext",
-    "module": "commonjs",
-    "declaration": true,
-    "outDir": "lib",
-    "allowSyntheticDefaultImports": true,
-    "esModuleInterop": true,
-    "sourceMap": true,
-    "jsx": "react",
-    "moduleResolution": "node",
-    "emitDecoratorMetadata": true,
-    "experimentalDecorators": true,
-    "noImplicitAny": false,
-    "noImplicitReturns": false,
-    "removeComments": true,
-    "strictNullChecks": false,
-    "skipLibCheck": true,
-    "resolveJsonModule": true,
-    "lib": [
-      "es6",
-      "es7",
-      "dom"
-    ],
-    "allowJs": false
-  }
-}
